% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archiveRenv.R
\name{archiveRenv}
\alias{archiveRenv}
\title{archiveRenv}
\usage{
archiveRenv()
}
\value{
Invisibly, the absolute path to the created lockfile.
}
\description{
Writes a new <timestamp>_renv.lock based on the current package environment into renv::project()/renv/archive.
}
\author{
Pascal Sauer
}
