% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkDeps.R
\name{checkDeps}
\alias{checkDeps}
\title{Check Dependencies}
\usage{
checkDeps(
  descriptionFile = ".",
  dependencyTypes = c("Depends", "Imports", "LinkingTo"),
  action = "stop"
)
}
\arguments{
\item{descriptionFile}{Path to a DESCRIPTION file or a path that belongs to a
source package project.}

\item{dependencyTypes}{The types of dependencies to check. Must be a
subset of \code{c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances")}.}

\item{action}{Action to take on unmet dependencies:
\itemize{
\item \code{"stop"}:  Issue an error with the unmet dependencies.  (Default.)
\item \code{"warn"}:  Issue a warning with the unmet dependencies.
\item \code{"pass"}:  Do nothing, just return invisibly.
\item \code{"ask"}: Ask the user whether to auto-fix missing dependencies. Requires an active renv.
}}
}
\value{
Invisibly, a named list of strings indicating whether each package
requirement is met (\code{"TRUE"}) or not, in which case the reason is stated.
}
\description{
Check if package requirements specified in the given DESCRIPTION are met.
}
\examples{
checkDeps(system.file("DESCRIPTION", package = "piamenv"))

}
\author{
Pascal Führlich, Michaja Pehl
}
