% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{checkPythonDeps}
\alias{checkPythonDeps}
\title{Check Python Dependencies and Optionally Their Versions}
\usage{
checkPythonDeps(dependencies, action = "stop", strict = TRUE)
}
\arguments{
\item{dependencies}{Vector of dependency strings, e.g. c("numpy==1.26.4", "climate-assessment @ <URL to git repo>")}

\item{action}{Action to take if a dependency is missing. Either "stop", "warn", "note", or "pass"}

\item{strict}{Logical indicating whether to check for matching versions}
}
\value{
TRUE if all dependencies are installed, FALSE otherwise
}
\description{
Checks if the required Python dependencies can actually be imported in the Python environment provided. Relies on
reticulate::import. Optionally check if versions are met
}
\examples{
\dontrun{
deps <- c("climate_assessment==0.1.4a0", "numpy<2.0")
checkPythonDeps(deps, action="stop", strict = TRUE)
}
}
