% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonEnv.R
\name{checkPythonEnv}
\alias{checkPythonEnv}
\title{checkPythonEnv.R}
\usage{
checkPythonEnv(path)
}
\arguments{
\item{path}{Optional. Fully qualified path to the Python environment}
}
\description{
Determine the host system's Python installation
}
\author{
Tonn Rüter
}
