% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{checkPythonRequirements}
\alias{checkPythonRequirements}
\title{Check Python Dependencies and Versions}
\usage{
checkPythonRequirements(
  requirementsFile,
  installed = py_list_packages(),
  action = "stop",
  strict = TRUE
)
}
\arguments{
\item{requirementsFile}{Path to a file containing the Python requirements. pip-style requirements are expected}

\item{installed}{Python packages installed in the environment. Defaults to reticulate::py_list_packages()}

\item{action}{Action if a dependency or version is missing or mismatched. Either "stop", "warn", "note", or "pass"}

\item{strict}{Logical indicating whether to check for strict version matching. Default is TRUE}
}
\value{
Vector of missing packages.
}
\description{
Compares requirements given in a pip-style file with the installed Python packages. Optionally check if versions are
met.
}
