% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{comparePythonVersions}
\alias{comparePythonVersions}
\title{Compare Python Versions}
\usage{
comparePythonVersions(operator, first, second, strict = TRUE)
}
\arguments{
\item{operator}{Comparison operator. One of '==', '!=', '>', '>=', '<' or '<='}

\item{first}{First Python version (represented as named list)}

\item{second}{Second Python version (represented as named list) to compare against}

\item{strict}{Strict comparison of version components (e.g. 3.10 != 3.10.0)}
}
\value{
TRUE operator is satisfied, FALSE otherwise
}
\description{
This function compares two Python versions
}
