% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condaInit.R
\name{condaInit}
\alias{condaInit}
\title{condaInit}
\usage{
condaInit(how = "", log = NULL, verbose = 0)
}
\arguments{
\item{how}{Method to activate conda if not found (e.g., "pik-cluster")}

\item{log}{Path to log file or NULL to use stdout}

\item{verbose}{Verbosity level (0 for silent, higher for more verbose)}
}
\value{
List containing `setup` and `teardown` commands, and environment variables (`env`).
}
\description{
Determine initialization commands to make conda available
}
