% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condaRun.R
\name{condaRun}
\alias{condaRun}
\title{condaRun}
\usage{
condaRun(
  cmd,
  path,
  env = c(),
  init = list(setup = "", teardown = ""),
  log = NULL,
  verbose = 0
)
}
\arguments{
\item{cmd}{Command to run in the conda environment}

\item{path}{Path to the conda environment}

\item{env}{Named vector of environment variables to set, i.e., c("VAR0" = "value0", "VAR1" = "value1")}

\item{init}{List containing `setup` and `teardown` commands to run before and after the command}

\item{log}{Path to log file or NULL to use stdout}

\item{verbose}{Verbosity level (0 for silent, higher for more verbose).}
}
\value{
Invisible NULL
}
\description{
Run a command in a specified conda environment with given environment variables.
}
