% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{createPythonDependency}
\alias{createPythonDependency}
\title{Create a named list representation of a Python dependency}
\usage{
createPythonDependency(
  name = "",
  operator = "",
  version = createPythonVersion(),
  build = "",
  repo = ""
)
}
\arguments{
\item{name}{Name of the dependency}

\item{operator}{Comparison operator. One of '==', '!=', '>', '>=', '<' or '<='}

\item{version}{Named list representation of the Python version of the dependency}

\item{build}{Build version of the dependency}

\item{repo}{Repository URL of the dependency}
}
\description{
This function creates a named list representing a dependency on a particular Python package
}
