% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{createPythonVersion}
\alias{createPythonVersion}
\title{checkPythonDeps.R}
\usage{
createPythonVersion(
  major = "",
  minor = "",
  patch = "",
  releaseType = "",
  releaseVersion = ""
)
}
\arguments{
\item{major}{Major version number}

\item{minor}{Minor version number}

\item{patch}{Patch version number}

\item{releaseType}{Release type (e.g. alpha, beta, rc)}

\item{releaseVersion}{Release version number}
}
\value{
A named list representing a Python version string
}
\description{
Verify the availability of REMIND's Python dependencies on the host system. Exposes two functions:
\code{checkPythonRequirements} and \code{checkPythonDeps}. While \code{checkPythonRequirements} reads a pip-style
requirements file and compares the required packages with the installed ones, \code{checkPythonDeps} checks if the
required Python dependencies can actually be imported in the Python environment provided.
}
\author{
Tonn Rüter
Create Python Version

This function creates a named list representing a Python version string from its components.
}
