% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResultsfolderPythonVirtualEnv.R
\name{createResultsfolderPythonVirtualEnv}
\alias{createResultsfolderPythonVirtualEnv}
\title{createResultsfolderPythonVirtualEnv}
\usage{
createResultsfolderPythonVirtualEnv(resultsfolder, sourceVenv = ".venv")
}
\arguments{
\item{resultsfolder}{Path to the resultsfolder where the new virtual environment will be created.}

\item{sourceVenv}{Path to the virtual environment folder which will be copied.}
}
\description{
Create a copy of the source virtual environment in the given results folder in the subfolder `.venv`.
}
\author{
Mika Pflüger
}
