% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{extractPythonDependency}
\alias{extractPythonDependency}
\title{Extract Python dependency from dependency string}
\usage{
extractPythonDependency(depString, style = "pip")
}
\arguments{
\item{depString}{Dependency string such as "numpy<=2.0"}

\item{style}{Style of the dependency string. Either "pip" or "conda"}
}
\value{
A named list representation of the Pzthon package name and version
}
\description{
This function extracts the package name, relationship operator and version from a typical Python dependency string
(e.g. "numpy==1.26.4") and returns a named list representing the dependency. Dependencies on specific repositories
are also supported (e.g. "climate-assessment @ <URL to git repo>")
}
