% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{extractPythonVersion}
\alias{extractPythonVersion}
\title{Extract Python Version}
\usage{
extractPythonVersion(versionString)
}
\arguments{
\item{versionString}{A Python version string like "3.10.0a0"}
}
\value{
A named list representing the Python version
}
\description{
This function extracts a named list representing Python version from a string by matching a regular expression.
}
