% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDeps.R
\name{fixDeps}
\alias{fixDeps}
\title{Fix Dependencies}
\usage{
fixDeps(
  ask = FALSE,
  requirementMet = checkDeps(action = if (ask) "note" else "pass")
)
}
\arguments{
\item{ask}{Whether to ask before fixing dependencies.}

\item{requirementMet}{The output of piamenv::checkDeps.}
}
\value{
Invisibly, the return value of renv::install.
}
\description{
Automatically install package versions as suggested by piamenv::checkDeps into an renv.
}
\author{
Pascal Führlich
}
