% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installDeps.R
\name{installDeps}
\alias{installDeps}
\title{installDeps}
\usage{
installDeps(path)
}
\arguments{
\item{path}{The path to scan for dependencies.}
}
\description{
Install dependencies, identified by renv::dependencies for a given path, into the current renv.
}
\author{
Pascal Sauer
}
