% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piamenv-package.R
\docType{package}
\name{piamenv-package}
\alias{piamenv}
\alias{piamenv-package}
\title{piamenv: Package environment support for PIAM}
\description{
Enables easier management of package environments, based on renv and conda.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/piamenv}
}

}
\author{
\strong{Maintainer}: Pascal Sauer \email{pascal.sauer@pik-potsdam.de}

Authors:
\itemize{
  \item Tonn Rüter \email{tonn.rueter@pik-potsdam.de}
}

}
