% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pythonBinPath.R
\name{pythonBinPath}
\alias{pythonBinPath}
\title{pythonBinPath}
\usage{
pythonBinPath(venv)
}
\arguments{
\item{venv}{Path to Python virtual environment folder.}
}
\value{
The full path to the Python binary in the virtual environment folder.
}
\description{
Returns the proper path to the Python binary in a given virtual environment.
}
\author{
Mika Pflüger
}
