% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{readPythonRequirements}
\alias{readPythonRequirements}
\title{checkPythonDeps.R}
\usage{
readPythonRequirements(path)
}
\arguments{
\item{path}{Path to the requirements file}
}
\value{
A sorted vector of dependencies
}
\description{
Verify the availability of REMIND's Python dependencies on the host system. Exposes two functions:
\code{checkPythonRequirements} and \code{checkPythonDeps}. While \code{checkPythonRequirements} reads a pip-style
requirements file and compares the required packages with the installed ones, \code{checkPythonDeps} checks if the
required Python dependencies can actually be imported in the Python environment provided.
}
\author{
Tonn Rüter
Read Python Requirements

This function reads a pip-style requirements file and returns a sorted vector of dependencies.
}
