% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{repoString}
\alias{repoString}
\title{Python dependency string with repository information}
\usage{
repoString(dependency)
}
\arguments{
\item{dependency}{Named list representation of the dependency}
}
\description{
This function creates a string representation of a Python dependency with repository information
}
