% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreRenv.R
\name{restoreRenv}
\alias{restoreRenv}
\title{restoreRenv}
\usage{
restoreRenv(
  lockfile = Sys.glob(c("output/*/renv.lock", "renv/archive/*renv.lock"))
)
}
\arguments{
\item{lockfile}{One or more paths to lockfiles. The default value assumes an output folder with
run folders containing a renv.lock file each, and/or an renv/archive folder with renv.lock files.}
}
\description{
Restores the current renv to the state described in the given lockfile.
If multiple lockfiles are given, ask the user which one to restore.
}
\author{
Pascal Führlich
}
