% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revertDevelopmentVersions.R
\name{revertDevelopmentVersions}
\alias{revertDevelopmentVersions}
\title{Revert Development Versions of Packages}
\usage{
revertDevelopmentVersions()
}
\value{
Invisibly the return value of \code{\link[renv:install]{renv::install()}}.
}
\description{
All PIK-PIAM packages in the current renv that are development versions, i.e.
that have a non-zero fourth version number component (e.g. \verb{0.4.3.9001}), are
reverted to the highest version lower than the development versions (e.g.
\verb{0.4.3}).
}
