% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showUpdates.R
\name{showUpdates}
\alias{showUpdates}
\title{showUpdates}
\usage{
showUpdates(packages = piamPackages())
}
\arguments{
\item{packages}{A character vector of package names.}
}
\value{
Invisibly, a data.frame as returned by utils::old.packages
}
\description{
Print available updates of the given packages.
}
