% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopIfLoaded.R
\name{stopIfLoaded}
\alias{stopIfLoaded}
\title{stopIfLoaded}
\usage{
stopIfLoaded(updatedPackage)
}
\arguments{
\item{updatedPackage}{One or more names of packages that were just updated.}
}
\description{
Throw an error if any of the given packages is loaded.
}
\details{
This is useful after updating packages. If any of the updated packages was loaded before
the update R might crash when trying to lazy load a function from the updated package.
}
\author{
Pascal Führlich
}
