% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updatePythonVirtualEnv.R
\name{updatePythonVirtualEnv}
\alias{updatePythonVirtualEnv}
\title{updatePythonVirtualEnv}
\usage{
updatePythonVirtualEnv(venv = ".venv", requirements = "requirements.txt")
}
\arguments{
\item{venv}{Path to Python virtual environment folder.}

\item{requirements}{Path to a requirements.txt file containing the current requirements.}
}
\description{
Installs newly added requirements into the Python virtual environment.
}
\author{
Mika Pflüger
}
