% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRenv.R
\name{updateRenv}
\alias{updateRenv}
\title{updateRenv}
\usage{
updateRenv(exclude = NULL)
}
\arguments{
\item{exclude}{vector of packages not to be updated}
}
\value{
Invisibly, the return value of renv::update.
}
\description{
Update all PIK-PIAM packages in the current renv, write renv.lock into archive.
}
\author{
Pascal Sauer
}
