% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writePythonVirtualEnvLockFile.R
\name{writePythonVirtualEnvLockFile}
\alias{writePythonVirtualEnvLockFile}
\title{writePythonVirtualEnvLockFile}
\usage{
writePythonVirtualEnvLockFile(venv = ".venv", filePath)
}
\arguments{
\item{venv}{Path to Python virtual environment folder.}

\item{filePath}{Path where the virtual environment lock file will be created.}
}
\description{
Write Python virtual environment lock file specifying package versions currently installed in the given
virtual environment. Using the lock file, a new virtual environment can be built containing exactly the same
package versions.
}
\author{
Mika Pflüger
}
