% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPythonDeps.R
\name{createPythonVersion}
\alias{createPythonVersion}
\title{Create Python Version}
\usage{
createPythonVersion(
  major = "",
  minor = "",
  patch = "",
  releaseType = "",
  releaseVersion = ""
)
}
\arguments{
\item{major}{Major version number}

\item{minor}{Minor version number}

\item{patch}{Patch version number}

\item{releaseType}{Release type (e.g. alpha, beta, rc)}

\item{releaseVersion}{Release version number}
}
\value{
A named list representing a Python version string
}
\description{
This function creates a named list representing a Python version string from its components.
}
