% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restoreRenv.R
\name{restoreRenv}
\alias{restoreRenv}
\title{restoreRenv}
\usage{
restoreRenv(lockfile)
}
\arguments{
\item{lockfile}{One or more paths to lockfiles.}
}
\description{
Restores the current renv to the state described in the given lockfile.
If multiple lockfiles are given, ask the user which one to restore.
}
\author{
Pascal Führlich
}
