% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showUpdates.R
\name{showUpdates}
\alias{showUpdates}
\title{showUpdates}
\usage{
showUpdates(packages = NULL)
}
\arguments{
\item{packages}{A character vector of package names. By default the result of
lucode2::pikPiamPackages() if lucode2 is available, otherwise must be passed explicitly.}
}
\value{
Invisibly, a data.frame as returned by utils::old.packages
}
\description{
Print available updates of the given packages.
}
